correct

+++ src/components/Objects/New/FormTemplate/FormBuilding.jsx
@@ -60,7 +60,7 @@ const FormBuilding = () => {
             name='BuildingTotalArea'
             control={control}
             render={({ field }) => (
-              <InputUI onChange={(e) => field.onChange(parseInt(e.target.value))}
+              <InputUI onChange={(e) => field.onChange(parseFloat(e.target.value))}
                 value={field.value || ''} label='Общая площадь здания' fullWidth type='number' />
             )}
           />
@@ -68,7 +68,7 @@ const FormBuilding = () => {
             name='LandArea'
             control={control}
             render={({ field }) => (
-              <InputUI onChange={(e) => field.onChange(parseInt(e.target.value))}
+              <InputUI onChange={(e) => field.onChange(parseFloat(e.target.value))}
                 value={field.value || ''} label='Площадь участка' fullWidth type='number' />
             )}
           />
