+++ src/components/Objects/New/FormTemplate/FormCommercialLand.jsx
@@ -48,7 +48,7 @@ const FormCommercialLand = () => {
             name='LandArea'
             control={control}
             render={({ field }) => (
-              <InputUI onChange={(e) => field.onChange(parseInt(e.target.value))}
+              <InputUI onChange={(e) => field.onChange(parseFloat(e.target.value))}
                 value={field.value || ''} label='Площадь участка' fullWidth type='number' />
             )}
           />
